Remove isMultiWriterCluster flag and ignore stale records when querying the topology (#435)
### Summary

Remove the isMultiWriterCluster flag from the AuroraTopologyService and
add checks for stale topology records.

### Description

Multi writer clusters are not longer supported, so the
isMultiWriterCluster flag that was used to disable failover has been
removed. Additionally, checks have been added to ensure stale writer
records that are obtained post writer-failover are not being used.

### Additional Reviewers

<!-- Any additional reviewers -->

### By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.