Fix bug where topology query could execute indefinitely (#416)
The topology query is not monitored by the EFM plugin. If the user has
no socket timeout and the topology query is executed against a down
instance, it could execute indefinitely. This PR temporarily sets a
socket timeout for the topology query if none was set by the user.

### By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.