Don't get connection for closed statement
It is possible that .getConnectionFromSqlObject() is called for a statement that has already been closed. This is the case when e.g. calling .isClosed() for a closed Statement. In this case, an SQLException is thrown from statement.getConnection() unnecessarily. The exception might also be detected in various APM product instrumentations leading to false positive errors.