Don't get Statement from closed ResultSet
It's possible that getConnectionFromSqlObject is called on a ResultSet that is already closed,
e.g. in DefaultConnectionPlugin when calling .close on a ResultSet. With some implementations,
such as PgResultSet, calling getStatement on a closed ResultSet throws an SQLException. There is a
performance overhead associated with creating and catching these exceptions everytime a ResultSet is
used, and closed, so it should be avoided.